# feat: introduce a LRU compiled style cache for the HTML formatter